fix: PostgreSQL tinyint 改为 smallint，兼容跨数据库类型

+++ src/main/resources/zephyr-zh-CN.sql
@@ -27,6 +27,7 @@ ALTER TABLE zephyr_knowledge_base ADD COLUMN IF NOT EXISTS scope varchar(16) DEF
 
 ALTER TABLE zephyr_model_configs ADD COLUMN IF NOT EXISTS scope VARCHAR(16) DEFAULT 'user';
 
-ALTER TABLE zephyr_knowledge_base ADD COLUMN IF NOT EXISTS graph_enabled SMALLINT DEFAULT 0;ALTER TABLE zephyr_knowledge_doc ADD COLUMN IF NOT EXISTS graph_status varchar(16);
+ALTER TABLE zephyr_knowledge_base ADD COLUMN IF NOT EXISTS graph_enabled SMALLINT DEFAULT 0;
+ALTER TABLE zephyr_knowledge_doc ADD COLUMN IF NOT EXISTS graph_status varchar(16);
 
-alter table if exists zephyr_workspaces add column if not exists is_system tinyint default 0;
+alter table if exists zephyr_workspaces add column if not exists is_system smallint default 0;
